backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
anthropic==0.21.3
sentence-transformers==2.6.1
faiss-cpu==1.8.0
pypdf2==3.0.1
pydantic==2.6.3
numpy==1.26.4
scikit-learn==1.4.1.post1
slowapi==0.1.9
python-jose==3.3.0
aiofiles==23.2.1
httpx==0.27.0
tenacity==8.2.3
python-dotenv==1.0.1
